Revert "Chat"
