+++ lib/src/bloc/auth/auth_state.dart
@@ -14,7 +14,3 @@ class Authenticated extends AuthState {}
 class UnAuthenticated extends AuthState {}
 
 class Loading extends AuthState {}
-
-class Home extends AuthState {}
-
-class Chat extends AuthState {}
