--- lib/src/model/channel.dart
@@ -1,10 +1,0 @@
-class Channel {
-  String id;
-  String name;
-  List<String> members;
-  String type;
-  String dp;
-  List<String> admins;
-
-  Channel({this.id, this.name, this.members, this.type, this.dp, this.admins});
-}
