+++ lib/src/ui/Registeration_screen 2.dart
@@ -3,9 +3,6 @@ import 'package:flutter/material.dart';
 import 'Pin_setup.dart';
 
 class PinSetup extends StatefulWidget {
-  final String text;
-  PinSetup(this.text);
-
   @override
   _PinSetupState createState() => _PinSetupState();
 }
@@ -67,11 +64,9 @@ class _PinSetupState extends State<PinSetup> {
                       size: 30.0,
                     ),
                     onPressed: () {
-                      print(widget.text);
                       Navigator.push(
                         context,
-                        MaterialPageRoute(
-                            builder: (context) => PinPage(widget.text)),
+                        MaterialPageRoute(builder: (context) => PinPage()),
                       );
                     })
               ],
